seperating components
